Supply_Chain_Revenue_Forecasting_Pipeline/pyproject.toml
[project]
name = "supply-chain-revenue-forecasting-pipeline"
version = "1"
description = "This project implements a modular, end-to-end data pipeline to forecast supply chain revenue using Airflow, PostgreSQL, and machine learning. The final output is a CSV file suitable for Tableau dashboards."
authors = [
    {name = "Hrishikesh"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = "3.9"
dependencies = [
    "pandas>=2.0",
    "scikit-learn>=1.0",
    "sqlalchemy>=2.0",
    "psycopg2-binary>=2.9",
    "joblib>=1.3",
    "apache-airflow>=2.7"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
